PageSize as uint
git-svn-id: https://svn.code.sf.net/p/flyhigh/code/trunk@1435 1728cf9f-2348-4ec0-968f-f422e243bc8e

+++ src/database/flytec/Defs5020.h
@@ -61,7 +61,7 @@
 #define BATT_TYPE_POS 352
 #define UTC_HALF_OFFSET_POS 384
 
-static const uchar Flytec5020PageSize = 8;
+static const uint Flytec5020PageSize = 8;
 
 static const uint Flytec5020MemSize = 385;
